+++ src/controllers/index.ts
@@ -1,7 +1,9 @@
+import { File } from './file'
 import { List } from './list'
 import { User } from './user'
 
 export {
+  File,
   List,
   User
 }