+++ src/controllers/utils/file.messages.ts
@@ -1,0 +1,5 @@
+enum ErrorForFile {
+  formatNotAllowed = 'Solo el está permitido el formato pdf.'
+}
+
+export { ErrorForFile as EFF }