+++ src/database/index.ts
@@ -1,4 +1,5 @@
 import { firebaseConnection } from './firebase'
+import { mongoConnection } from './mongoose'
 import { redisConnection } from './redis'
 
-export { firebaseConnection, redisConnection }
+export { firebaseConnection, mongoConnection, redisConnection }