+++ src/dto-interfaces/file.dto.ts
@@ -1,0 +1,8 @@
+export interface DtoFile {
+  data    : Buffer
+  encoding: string
+  list    : string
+  mimetype: string
+  name    : string
+  size    : number
+}