+++ src/dto-interfaces/index.ts
@@ -1,4 +1,5 @@
+import { DtoFile } from './file.dto'
 import { DtoList } from './list.dto'
 import { DtoUser } from './user.dto'
 
-export { DtoList, DtoUser }
+export { DtoFile, DtoList, DtoUser }