+++ src/models/index.ts
@@ -1,0 +1,6 @@
+import { IFile, FileModel } from './files'
+
+export {
+  IFile,
+  FileModel
+}