+++ src/routes/index.ts
@@ -1,10 +1,12 @@
 import { Auth } from './auth'
+import { File } from './file'
 import { Home } from './home'
 import { List } from './list'
 import { User } from './user'
 
 export {
   Auth,
+  File,
   Home,
   List,
   User